# OpenAPI.yaml
openapi: 3.0.0
info:
  title: store
  version: '1.0'
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  termsOfService: 'https://www.securitease.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'
servers:
  - url: 'http://localhost:8080'
    description: dev
paths:
  /order:
    get:
      summary: Get Orders
      tags:
        - store
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
            X-Total-Pages:
              $ref: '#/components/headers/XTotalPages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
      operationId: getOrders
      description: >-
        Get all orders. Omit both `page` and `size` to receive the full,
        unpaged list (existing behaviour, kept for backwards compatibility).
        Supplying either parameter switches to a bounded page and adds the
        `X-Total-Count`/`X-Total-Pages` response headers.
    post:
      summary: Create Order
      tags:
        - store
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      operationId: createOrder
      description: Create a new order
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                descriptioon:
                  type: string
  /order/{id}:
    get:
      summary: Get Order By Id
      tags:
        - store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
      operationId: getOrderById
      description: Get a specific order by its ID
  /customer:
    get:
      summary: Get Customers
      tags:
        - store
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: >-
            Filter to customers with a word in their name containing this
            substring (case-insensitive). Omit or leave blank for all customers.
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
            X-Total-Pages:
              $ref: '#/components/headers/XTotalPages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
      operationId: getCustomers
      description: >-
        Get all customers, optionally filtered by `name`. Omit both `page`
        and `size` to receive the full, unpaged list (existing behaviour,
        kept for backwards compatibility). Supplying either parameter
        switches to a bounded page and adds the
        `X-Total-Count`/`X-Total-Pages` response headers.
    post:
      summary: Create Customer
      tags:
        - store
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
      operationId: createCustomer
      description: Create a new customer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                name:
                  type: string
  /products:
    get:
      summary: Get Products
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  description:
                    type: string
                  orders:
                    type: array
                    items:
                      type: integer
                      format: int64
      operationId: getProducts
      description: Get all products
    post:
      summary: Create Product
      tags:
        - store
      responses:
        '201':
          description: Created
      operationId: createProduct
      description: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                description:
                  type: string
                  x-stoplight:
                    id: jbh0jzkonppv8
components:
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
      description: >-
        Zero-based page number, defaulting to 0. Providing this parameter
        (with or without `size`) opts into paging.
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
      description: >-
        Page size, capped at a server-enforced maximum. Providing this
        parameter (with or without `page`) opts into paging.
  headers:
    XTotalCount:
      description: Total number of items matching the request. Present only when paging is requested.
      schema:
        type: integer
        format: int64
    XTotalPages:
      description: Total number of pages. Present only when paging is requested.
      schema:
        type: integer
        format: int32
  schemas:
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          type: object
          properties:
            id:
              type: integer
              format: int64
            description:
              type: string
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string
                x-stoplight:
                  id: mq08x82g95zl9
tags:
  - name: store
    description: store operations